Move code inside setup
Move code inside setup because it turns out the
Robot global isn't defined outside of the setup
and loop functions.
Flip motors because we changed which side of the
robot was the front.
Change motor gear ratio to 1:30 since that seemed
to fix tests at first. (Great solution, I know.)